Push and retrieve numbers from the stack
It is now possible to push and retrieve Lua Numbers (floating point) to/
from the stack.

The Lua library can be built with different C types underneath its
Number and Integer types. For now, I am leaving the C types as
automatically declared based on lua.h (e.g. lua_Number) in Lua.Internal
and defining my own Ada-based version of Number in Lua. Explicit type
conversions are needed internally but should disappear on compilation on
many platforms (e.g. GCC/GNAT x86_64m) where the standard Ada types are
the same as the corresponding C types. Down the road, this should allow
the bindings to be usable with non-standard configurations of Lua.