Added Lua_Reference type
Persistent references to Lua values can now be held via the
Lua_Reference type. This is itself a reference-counted type using the
Ada.Finalization facilities that ensures that the Lua reference is
'unref'ed as appropriate. Reference-counting on top of a mark-sweep
garbage collector...

Currently this does not check for the case where a reference outlives
the lifetime of the underlying State.