Add_Yield_Function as alternative to coroutine lib
The Add_Yield_Function procedure in Lua.Libs can be used to add a
straightforward yield function to a Lua state without needing to import
the coroutine library in its entirety. This may be useful where users
are not expected to set up their own threads or coroutines, and having
to use "coroutine.yield" to yield might seem strange.