Add support for Ada tagged types as Lua Userdata
The Lua.Userdata generic package contains routines to allow a tagged
type to be used as a Userdata value in Lua. Operations can be added to
the Lua type's metatable so its values can be used in an object-oriented
fashion in Lua. Only tagged types are supported currently, as the T'Tag
is used as a sentinel to do some type checking when userdata is
retrieved. T'External_Tag is used as part of the name of the metatable
in the Lua registry.