Renamed deliberately unused variables to Discard
GNAT will complain about an assignment to a variable where the variable
is then not used for anything, unless the variable name is one of a few
'magic' values such as Discard. Where it is clear that the return value
does not need to be checked it is safe to change the name of the
variable.