Re-think the treatment of class-wide userdata
The treatment of userdata created from access-to-T'Class values has been
completely rethought, together with the approach to automatic
conversions. Conversions between userdata created from different types
is now allowed whenever the type used to create the actual userdata is
covered by the type of the requested userdata.

Note that this means that if C derives from P, a userdata created from a
P'Class(C_value) cannot convert to a userdata of type C or C'Class, even
though the conversion would be valid at run-time in Ada. In Lua the
distinction between userdata created from P that would or would not also
be valid as userdata from C is not observable, and so it would create
unpredictable behaviour.