Resume can take a null 'from' parameter
The Resume function for threads can now takes a (default) constant
Null_Thread for the 'from' parameter which corresponds to passing a
null pointed to the 'from' parameter of the C API call. The order of
parameters has been changed from the C API so the default parameter is
at the end.