Rename GC function to GC_IsRunning
In the C API the function 'lua_gc' is overloaded depending on the value
of an 'op' parameter. In the Ada API these are split out. However there
is only one valid version that returns a Boolean, so it would be better
to change the name of this to give a better indication of what it is
indicating.