Stream_Element'Size might not always be 8
Correct Stream_Lua_Writer to account for cases in which
Stream_Element'Size is not 8 bits but it is a integral multiple of 8
bits.

If there are any Ada systems out there where Stream_Element'Size is not
a multiple of 8 then this code will not work, but to be honest I do not
think it is practical for such systems to inter-operate with a C library
such as the Lua interpreter.

RM13.13.1 9.1/1 seems to suggest that Stream_Element'Size does not even
have to be a multiple of Storage_Element'Size...