LoadString_By_Copy always converts to C convention
If a string containing code has to be copied in order to be passed to
the Lua interpreter, perhaps because the Ada and C conventions on the
system are so different, it is probably futile to try to avoid using
the conversion routine in Interfaces.C.Strings. Where the routine is
used for convenience it doesn't matter either way. It is safer to always
convert the string using the proper routine.