Improve IsAdaFunction by hiding trampoline
Specifications for routines only used by the package body of Lua are now
declared at the start of the package body. As CFunction_Trampoline is
now completely hidden, we can now test CFunctions on the stack to see
if they are AdaFunctions by simple comparison with
CFunction_Trampoline'Access without compromising safety.