Fix minor memory leak in Lua.Libs
The library names in the LibNames array were created using
Interfaces.C.Strings.New_String which involves an implicit allocation
on library elaboration that will never be deallocated. This is not a
major problem as it is a small amount of memory and will not grow over
time. However to avoid this the library names are now stored as
fixed-width strings, and only trimmed and converted to C-style strings
as they are required.

Issue identified by the gnatmem tool.